requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.8.0
email-validator==2.3.0
emails==0.6
flasgger==0.9.7.1
Flask==3.1.2
flask-cors==6.0.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==6.0.2
mailersend==2.0.0
MarkupSafe==3.0.2
mistune==3.1.4
more-itertools==10.8.0
packaging==25.0
premailer==3.10.0
psycopg2-binary==2.9.10
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.1
six==1.17.0
SQLAlchemy==2.0.43
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
